functions/lambda_finder/requirements.txt
boto3==1.16.11
botocore==1.19.11
jmespath==0.10.0
progressbar2==3.53.1
python-dateutil==2.8.1
python-utils==2.4.0
s3transfer==0.3.3
six==1.15.0
terminaltables==3.1.0
urllib3==1.25.11
